# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_pos-retail_go_backend_internal_types_request.ProductCreateRequest:
    properties:
      barcode:
        maxLength: 50
        type: string
      category_id:
        type: string
      company_id:
        type: string
      cost_price:
        type: number
      description:
        type: string
      name:
        maxLength: 200
        type: string
      reorder_point:
        minimum: 0
        type: integer
      retail_price:
        type: number
      sku:
        maxLength: 50
        type: string
      tax_rate:
        maximum: 100
        minimum: 0
        type: number
      unit_id:
        type: string
    required:
    - cost_price
    - name
    - retail_price
    - sku
    - unit_id
    type: object
  github_com_pos-retail_go_backend_internal_types_request.ProductUpdateRequest:
    properties:
      barcode:
        maxLength: 50
        type: string
      category_id:
        type: string
      cost_price:
        type: number
      description:
        type: string
      name:
        maxLength: 200
        type: string
      reorder_point:
        minimum: 0
        type: integer
      retail_price:
        type: number
      sku:
        maxLength: 50
        type: string
      tax_rate:
        maximum: 100
        minimum: 0
        type: number
      unit_id:
        type: string
    type: object
...
  github_com_pos-retail_go_backend_internal_types_response.ApiResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  github_com_pos-retail_go_backend_internal_types_response.PaginatedResponse:
    properties:
      data: {}
      pagination:
        $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.Pagination'
      success:
        type: boolean
    type: object
  github_com_pos-retail_go_backend_internal_types_response.Pagination:
    properties:
      has_more:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
info:
  contact: {}
  description: Go (Fiber) implementation of POS Retail Backend.
  title: POS Retail Backend API
  version: "1.0"
paths:
...
  /api/categories:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
      security:
      - BearerAuth: []
      summary: List categories
      tags:
      - Categories
...
  /api/price-tiers:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product ID
        in: query
        name: product_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
      security:
      - BearerAuth: []
      summary: List price tiers
      tags:
      - PriceTiers
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Price tier payload
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Create price tier
      tags:
      - PriceTiers
  /api/price-tiers/{id}:
    delete:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Price Tier ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Delete price tier
      tags:
      - PriceTiers
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Price Tier ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get price tier
      tags:
      - PriceTiers
    put:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Price Tier ID
        in: path
        name: id
        required: true
        type: string
      - description: Update payload
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Update price tier
      tags:
      - PriceTiers
  /api/products:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by category
        in: query
        name: category_id
        type: string
      - description: Search term
        in: query
        name: search
        type: string
      - default: 50
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.PaginatedResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
      security:
      - BearerAuth: []
      summary: List products
      tags:
      - Products
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_request.ProductCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Create product
      tags:
      - Products
  /api/products/{id}:
    delete:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Delete product
      tags:
      - Products
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get product
      tags:
      - Products
    put:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Product payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_request.ProductUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Update product
      tags:
      - Products
...
  /api/units:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
      security:
      - BearerAuth: []
      summary: List units
      tags:
      - Units
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"